adding wandb
