+++ scripts/train.py
@@ -1,38 +1,14 @@
 from __future__ import print_function
 import argparse
 import torch
-import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
 from torchvision import datasets, transforms
 from torch.optim.lr_scheduler import StepLR
-
-
-class Net(nn.Module):
-    def __init__(self):
-
-        super(Net, self).__init__()
-        self.conv1 = nn.Conv2d(1, 32, 3, 1)
-        self.conv2 = nn.Conv2d(32, 64, 3, 1)
-        self.dropout1 = nn.Dropout(0.25)
-        self.dropout2 = nn.Dropout(0.5)
-        self.fc1 = nn.Linear(9216, 128)
-        self.fc2 = nn.Linear(128, 10)
-
-    def forward(self, x):
-        x = self.conv1(x)
-        x = F.relu(x)
-        x = self.conv2(x)
-        x = F.relu(x)
-        x = F.max_pool2d(x, 2)
-        x = self.dropout1(x)
-        x = torch.flatten(x, 1)
-        x = self.fc1(x)
-        x = F.relu(x)
-        x = self.dropout2(x)
-        x = self.fc2(x)
-        output = F.log_softmax(x, dim=1)
-        return output
+from plr_exercise.models.cnn import Net
+from plr_exercise import PLR_ROOT_DIR
+import wandb
+import os
 
 
 def train(args, model, device, train_loader, optimizer, epoch):
@@ -55,6 +31,7 @@ def train(args, model, device, train_loader, optimizer, epoch):
                     loss.item(),
                 )
             )
+            wandb.log({"epoch": epoch, "train_loss": loss.item()})
             if args.dry_run:
                 break
 
@@ -80,6 +57,7 @@ def test(model, device, test_loader, epoch):
             test_loss, correct, len(test_loader.dataset), 100.0 * correct / len(test_loader.dataset)
         )
     )
+    wandb.log({"test_loss": test_loss, "epoch": epoch})
 
 
 def main():
@@ -106,6 +84,18 @@ def main():
     )
     parser.add_argument("--save-model", action="store_true", default=False, help="For Saving the current Model")
     args = parser.parse_args()
+
+    wandb.login()
+    os.makedirs(os.path.join(PLR_ROOT_DIR, "results"), exist_ok=True)
+    run = wandb.init(
+        dir=os.path.join(PLR_ROOT_DIR, "results"),
+        project="plr-project",
+        config=args,
+        settings=wandb.Settings(code_dir=PLR_ROOT_DIR),
+    )
+    include_fn = lambda path, root: path.endswith(".py") or path.endswith(".yaml")
+    run.log_code(name="source_files", root=PLR_ROOT_DIR, include_fn=include_fn)
+
     use_cuda = not args.no_cuda and torch.cuda.is_available()
 
     torch.manual_seed(args.seed)
@@ -139,6 +129,8 @@ def main():
 
     if args.save_model:
         torch.save(model.state_dict(), "mnist_cnn.pt")
+
+    wandb.finish()
 
 
 if __name__ == "__main__":
